final version

+++ app.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
-const date = require(__dirname+"/date.js");
 const mongoose = require("mongoose");
+const _ = require("lodash");
 
 const app = express();
 
@@ -9,7 +9,7 @@ app.set("view engine","ejs");
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(express.static("public"));
 
-mongoose.connect("mongodb://localhost:27017/todolist",{ useUnifiedTopology: true ,useNewUrlParser: true});
+mongoose.connect("mongodb+srv://admin-om:bba_admin123@cluster0.9xjs3.mongodb.net/todolistDB?retryWrites=true&w=majority",{ useUnifiedTopology: true ,useNewUrlParser: true});
 
 const itemsSchema = {
     name:String,
@@ -43,56 +43,89 @@ function insertIfEmpty(){
 
 const listSchema = {
     name:String,
-    item=[itemsSchema]
+    item:[itemsSchema]
 }
 
 const List = mongoose.model("List",listSchema);
 
 app.get("/",(req,res)=>{
-    let day = date.getdate();
     Item.find({},(err,items)=>{
         if(items.length == 0){
             insertIfEmpty();
             res.redirect("/");
         }else{
-            res.render('list',{listTittle:day,toDoList:items});
+            res.render('list',{listTittle:"Today",toDoList:items});
         }
     });
 });
 
 app.post('/',(req,res)=>{
     let itemName = req.body.item;
-    if(req.body.list=='Work'){
-        workList.push(item);
-        res.redirect("/work")
-    }else{
-        const item = new Item({
-            name:itemName,
-        });
+    let listName = req.body.list;
+
+    const item = new Item({
+        name:itemName,
+    });
+    if(listName=="Today"){
         item.save();
         res.redirect("/");
+    }else{
+        //req.body.list
+        List.findOne({name:listName},(err,foundList)=>{
+            foundList.item.push(item);
+            foundList.save();
+            console.log(listName+" Success");
+            res.redirect("/"+listName);
+        });
     }
 });
 
 app.post("/delete",(req,res)=>{
     const checkerItemId = req.body.checkbox;
-    Item.findByIdAndRemove(checkerItemId,(err)=>{
-        if(err){
-            console.log(err);
-        }else{
-            console.log("Removed "+ checkerItemId);
-        }
-    });
-    res.redirect("/");
+    const listTittle = req.body.listName;
+
+    if (listTittle=="Today"){
+        Item.findByIdAndRemove(checkerItemId,(err)=>{
+            if(err){
+                console.log(err);
+            }else{
+                console.log("Removed "+ checkerItemId);
+            }
+        });
+        res.redirect("/");
+    }else{
+        List.findOneAndUpdate({name:listTittle},{$pull:{item:{_id:checkerItemId}}},(err,result)=>{
+            if(!err){
+                console.log("Removed "+ checkerItemId);
+                res.redirect("/"+listTittle);
+            }
+        });
+    }
 });
 
 app.get("/:route",(req,res)=>{
-    const listNmae = req.params.route;
-    const list = new List({
-        name:listNmae,
+    const listName = _.capitalize(req.params.route);
+
+    List.findOne({name:listName},(err,foundList)=>{
+        if(!err){
+            if(!foundList){
+                //create list
+                const list = new List({
+                    name:listName,
+                    item:defaultItem,
+                });
+
+                list.save();
+                console.log(listName+" Success");
+                res.redirect("/"+listName);
+            }else{
+                //show list
+                res.render('list',{listTittle:listName,toDoList:foundList.item});
+            }
+        }
     });
 });
 
-app.listen(3000,()=>{
+app.listen(process.env.PORT || 3000,()=>{
     console.log("server running on port 3000");
 });
